requirements.txt
annotated-doc
annotated-types
anyio
bcrypt>=4.0.0,<5.0.0
certifi
cffi
charset-normalizer
click
cryptography
dnspython
ecdsa
email-validator
fastapi
h11
httptools
idna
passlib
pyasn1
pycparser
pydantic
pydantic-settings
pydantic_core
python-dotenv
python-jose
python-multipart
PyYAML
requests
rsa
six
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
psycopg2-binary
alembic
slowapi
redis
pytest
pytest-asyncio
pytest-cov
httpx
faker
psutil
boto3
botocore
resend
